+++ ricaun.Nuke/Components/ICompileExample.cs
@@ -1,6 +1,4 @@
 ﻿using Nuke.Common;
-using Nuke.Common.Utilities.Collections;
-using System.Linq;
 
 namespace ricaun.Nuke.Components
 {