+++ ricaun.Nuke/Components/IHazRelease.cs
@@ -1,14 +1,18 @@
-﻿using Nuke.Common;
+﻿using JetBrains.Annotations;
+using Nuke.Common;
 using Nuke.Common.IO;
 using Nuke.Common.ProjectModel;
+using Nuke.Common.Utilities.Collections;
 using ricaun.Nuke.Extensions;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace ricaun.Nuke.Components
 {
     /// <summary>
     /// IHazRelease
     /// </summary>
-    public interface IHazRelease : IHazMainProject, IHazSolution, INukeBuild
+    public interface IHazRelease : IHazMainProject, IHazSolution, IHazSign, INukeBuild
     {
         /// <summary>
         /// Folder Release 
@@ -71,6 +75,61 @@ namespace ricaun.Nuke.Components
             ZipExtension.CreateFromDirectory(sourceDirectory, zipFile, includeBaseDirectory);
 
             return fileName;
+        }
+
+        /// <summary>
+        /// BuildProjectsAndRelease
+        /// </summary>
+        /// <param name="projects"></param>
+        /// <param name="releaseProjectFiles"></param>
+        /// <param name="releasePackages"></param>
+        /// <param name="signProjects"></param>
+        public void BuildProjectsAndRelease(IEnumerable<Project> projects,
+            bool releaseProjectFiles = true,
+            bool releasePackages = true,
+            bool signProjects = true)
+        {
+            foreach (var project in projects)
+            {
+                Solution.BuildProject(project, (project) =>
+                {
+                    project.ShowInformation();
+
+                    if (signProjects) SignProject(project);
+
+                    var releaseDirectory = project.Directory / "bin" / "Release";
+                    var fileName = project.Name;
+                    var version = project.GetInformationalVersion();
+
+                    if (releaseProjectFiles)
+                    {
+                        var releaseFileName = CreateReleaseFromDirectory(releaseDirectory, fileName, version);
+                        Serilog.Log.Information($"Release: {releaseFileName}");
+                    }
+
+                    if (releasePackages)
+                    {
+                        Globbing.GlobFiles(releaseDirectory, "**/*.*nupkg")
+                            .ForEach(file =>
+                            {
+                                Serilog.Log.Information($"Copy nupkg: {file} to {ReleaseDirectory}");
+                                AbsolutePathExtensions.CopyToDirectory(file, ReleaseDirectory, ExistsPolicy.MergeAndOverwriteIfNewer);
+                            });
+                    }
+                });
+            }
+        }
+
+        /// <summary>
+        /// ReportSummaryProjectNames
+        /// </summary>
+        /// <param name="projects"></param>
+        public void ReportSummaryProjectNames(IEnumerable<Project> projects)
+        {
+            var names = string.Join(" | ", projects.Select(e => e.Name).ToArray());
+            ReportSummary(_ => _
+                .AddPair("Projects", names)
+            );
         }
     }
 }